# Make `--doctool` imply `--headless` to speed up class reference generation